+++ getaround_utils/lib/getaround_utils/utils/config_url.rb
@@ -9,15 +9,19 @@ module GetaroundUtils::Utils; end
 module GetaroundUtils::Utils::ConfigUrl
   def self.from_env(config_name, ...)
     env_url = ENV.fetch("#{config_name}_URL", ...)
+    env_usr = ENV.fetch("#{config_name}_USERNAME", nil)
     env_pwd = ENV.fetch("#{config_name}_PASSWORD", nil)
-    return if env_url.nil?
+    return if env_url.nil? || env_url.empty? # rubocop:disable Rails/Blank
 
     ::URI.parse(env_url).tap do |uri|
-      if env_pwd && !uri.userinfo
-        uri.userinfo = ":#{env_pwd}"
-      elsif env_pwd
-        uri.password ||= env_pwd
-      end
+      uri_usr, uri_pwd = uri.userinfo&.split(':', 2)
+      uri_pwd = uri_pwd.to_s if uri.userinfo&.include?(':')
+      usr = env_usr.nil? ? uri_usr : env_usr
+      pwd = env_pwd.nil? ? uri_pwd : env_pwd
+      userinfo = usr.to_s
+      userinfo += ":#{pwd}" unless pwd.nil?
+
+      uri.userinfo = userinfo unless userinfo.empty?
     end
   end
 end