+++ getaround_utils/spec/getaround_utils/utils/config_url_spec.rb
@@ -3,113 +3,315 @@
 require 'spec_helper'
 
 RSpec.describe GetaroundUtils::Utils::ConfigUrl do
+  # rubocop:disable RSpec/NestedGroups
   describe '.from_env' do
     subject { described_class.from_env(config_name) }
 
+    let(:config_name) { 'TEST_CONFIG' }
+
+    let(:url_key) { "#{config_name}_URL" }
+    let(:usr_key) { "#{config_name}_USERNAME" }
+    let(:pwd_key) { "#{config_name}_PASSWORD" }
+
     let(:mocked_env) do
-      {
-        'TEST_NO_USERNAME_URL' => "mysql://localhost:666/test",
-        'TEST_NO_USERNAME_PASSWORD' => "used-pwd",
-        'TEST_URL_PWD_URL' => "mysql://user1:my-pwd@localhost:666/test",
-        'TEST_URL_PWD_PASSWORD' => "not-used-because-present-in-url",
-        'TEST_BLANK_PWD_URL' => "mysql://user2:@localhost:666/test",
-        'TEST_BLANK_PWD_PASSWORD' => "not-used-because-present-in-url-even-if-blank",
-        'TEST_COMPUTE_URL' => "mysql://user3@localhost:666/test",
-        'TEST_COMPUTE_PASSWORD' => "must-be-used",
-        'TEST_NO_PWD_URL' => "mysql://user4@localhost:666/test",
-        'TEST_NO_AUTH_URL' => "mysql://localhost:666/test",
-        'TEST_NO_HOST_NO_AUTH_URL' => "sqlite:///db/test.db",
-        'TEST_NO_HOST_URL' => "sqlite://user5@/db/test.db",
-        'TEST_NO_HOST_PASSWORD' => "pwd",
-      }
+      {}
     end
 
     before do
       allow(ENV).to receive(:fetch).and_call_original
-      allow(ENV).to receive(:fetch).with("#{config_name}_URL", any_args).and_wrap_original do |_, *args, &blk|
+      allow(ENV).to receive(:fetch).with(url_key, any_args).and_wrap_original do |_, *args, &blk|
         mocked_env.fetch(*args, &blk)
       end
-      allow(ENV).to receive(:fetch).with("#{config_name}_PASSWORD", any_args).and_wrap_original do |_, *args, &blk|
+      allow(ENV).to receive(:fetch).with(usr_key, any_args).and_wrap_original do |_, *args, &blk|
         mocked_env.fetch(*args, &blk)
+      end
+      allow(ENV).to receive(:fetch).with(pwd_key, any_args).and_wrap_original do |_, *args, &blk|
+        mocked_env.fetch(*args, &blk)
+      end
+    end
+
+    context 'with empty userinfo in base url' do
+      let(:mocked_env) do
+        {
+          url_key => 'mysql://@localhost:666/test',
+        }
+      end
+
+      it 'cleanup the base url' do
+        expect(subject).to be_a ::URI::Generic
+        expect(subject.to_s).to eq 'mysql://localhost:666/test'
+      end
+
+      context 'with usr variable matching' do
+        let(:mocked_env) do
+          super().merge(usr_key => 'user42')
+        end
+
+        it 'computes it without adding pwd' do
+          expect(subject).to be_a ::URI::Generic
+          expect(subject.to_s).to eq 'mysql://user42@localhost:666/test'
+        end
+      end
+
+      context 'with pwd variable matching' do
+        let(:mocked_env) do
+          super().merge(pwd_key => 'my-pwd')
+        end
+
+        it 'computes it correctly' do
+          expect(subject).to be_a ::URI::Generic
+          expect(subject.to_s).to eq 'mysql://:my-pwd@localhost:666/test'
+        end
+      end
+
+      context 'with usr + pwd variables matching' do
+        let(:mocked_env) do
+          super().merge(usr_key => 'user42', pwd_key => 'my-pwd')
+        end
+
+        it 'computes it' do
+          expect(subject).to be_a ::URI::Generic
+          expect(subject.to_s).to eq 'mysql://user42:my-pwd@localhost:666/test'
+        end
+      end
+    end
+
+    context 'with blank credentials in base url' do
+      let(:mocked_env) do
+        {
+          url_key => 'mysql://:@localhost:666/test',
+        }
+      end
+
+      it 'keeps them blank' do
+        expect(subject).to be_a ::URI::Generic
+        expect(subject.to_s).to eq 'mysql://:@localhost:666/test'
+      end
+
+      context 'with usr variable matching' do
+        let(:mocked_env) do
+          super().merge(usr_key => 'user42')
+        end
+
+        it 'computes it keeping blank pwd' do
+          expect(subject).to be_a ::URI::Generic
+          expect(subject.to_s).to eq 'mysql://user42:@localhost:666/test'
+        end
+      end
+
+      context 'with pwd variable matching' do
+        let(:mocked_env) do
+          super().merge(pwd_key => 'my-pwd')
+        end
+
+        it 'computes it correctly' do
+          expect(subject).to be_a ::URI::Generic
+          expect(subject.to_s).to eq 'mysql://:my-pwd@localhost:666/test'
+        end
+      end
+
+      context 'with usr + pwd variables matching' do
+        let(:mocked_env) do
+          super().merge(usr_key => 'user42', pwd_key => 'my-pwd')
+        end
+
+        it 'computes it' do
+          expect(subject).to be_a ::URI::Generic
+          expect(subject.to_s).to eq 'mysql://user42:my-pwd@localhost:666/test'
+        end
+      end
+    end
+
+    context 'with a usr variable and no usr in base url' do
+      let(:mocked_env) do
+        {
+          url_key => 'mysql://localhost:666/test',
+          usr_key => 'user42',
+        }
+      end
+
+      it 'computes the usr in the base url' do
+        expect(subject).to be_a ::URI::Generic
+        expect(subject.to_s).to eq 'mysql://user42@localhost:666/test'
+      end
+    end
+
+    context 'with a usr variable matching and pwd already present in base url' do
+      let(:mocked_env) do
+        {
+          url_key => 'mysql://:my-pwd@localhost:666/test',
+          usr_key => 'user42',
+        }
+      end
+
+      it 'computes the usr in the base url' do
+        expect(subject).to be_a ::URI::Generic
+        expect(subject.to_s).to eq 'mysql://user42:my-pwd@localhost:666/test'
+      end
+    end
+
+    context 'with a usr variable matching and usr + pwd already present in base url' do
+      let(:mocked_env) do
+        {
+          url_key => 'mysql://usr666:my-pwd@localhost:666/test',
+          usr_key => 'user42',
+        }
+      end
+
+      it 'overrides the usr in the base url' do
+        expect(subject).to be_a ::URI::Generic
+        expect(subject.to_s).to eq 'mysql://user42:my-pwd@localhost:666/test'
+      end
+    end
+
+    context 'with a usr + pwd variable matching and usr + pwd already present in base url' do
+      let(:mocked_env) do
+        {
+          url_key => 'mysql://usr666:my-pwd@localhost:666/test',
+          usr_key => 'user42',
+          pwd_key => 'new-pwd',
+        }
+      end
+
+      it 'overrides the usr in the base url' do
+        expect(subject).to be_a ::URI::Generic
+        expect(subject.to_s).to eq 'mysql://user42:new-pwd@localhost:666/test'
       end
     end
 
     context 'with pwd variable but no username in base url' do
-      let(:config_name) { 'TEST_NO_USERNAME' }
+      let(:mocked_env) do
+        {
+          url_key => 'mysql://localhost:666/test',
+          pwd_key => 'used-pwd',
+        }
+      end
 
       it 'computes the pwd in the base url' do
         expect(subject).to be_a ::URI::Generic
-        expect(subject.to_s).to eq "mysql://:used-pwd@localhost:666/test"
+        expect(subject.to_s).to eq 'mysql://:used-pwd@localhost:666/test'
       end
     end
 
-    context 'when pwd is already present in base url' do
-      let(:config_name) { 'TEST_URL_PWD' }
+    context 'when credentials are already present in base url' do
+      let(:mocked_env) do
+        {
+          url_key => 'mysql://user1:my-pwd@localhost:666/test',
+        }
+      end
 
-      it 'does not replace the url password' do
-        expect(subject).to be_a ::URI::Generic
-        expect(subject.to_s).to eq "mysql://user1:my-pwd@localhost:666/test"
+      context 'with pwd variable matching' do
+        let(:mocked_env) do
+          super().merge(pwd_key => 'override-url-pwd')
+        end
+
+        it 'overrides the url password' do
+          expect(subject).to be_a ::URI::Generic
+          expect(subject.to_s).to eq 'mysql://user1:override-url-pwd@localhost:666/test'
+        end
+      end
+
+      context 'with usr variable matching' do
+        let(:mocked_env) do
+          super().merge(usr_key => 'user42')
+        end
+
+        it 'overrides the url username' do
+          expect(subject).to be_a ::URI::Generic
+          expect(subject.to_s).to eq 'mysql://user42:my-pwd@localhost:666/test'
+        end
       end
     end
 
     context 'when pwd is blank in base url' do
-      let(:config_name) { 'TEST_BLANK_PWD' }
+      let(:mocked_env) do
+        {
+          url_key => 'mysql://user2:@localhost:666/test',
+          pwd_key => 'override-blank-pwd',
+        }
+      end
 
-      it 'does not replace the url password' do
+      it 'overrides the blank password' do
         expect(subject).to be_a ::URI::Generic
-        expect(subject.to_s).to eq "mysql://user2:@localhost:666/test"
+        expect(subject.to_s).to eq 'mysql://user2:override-blank-pwd@localhost:666/test'
       end
     end
 
     context 'with no pwd in base url and pwd variable' do
-      let(:config_name) { 'TEST_COMPUTE' }
+      let(:mocked_env) do
+        {
+          url_key => 'mysql://user3@localhost:666/test',
+          pwd_key => 'must-be-used',
+        }
+      end
 
       it 'computes the pwd in the base url' do
         expect(subject).to be_a ::URI::Generic
-        expect(subject.to_s).to eq "mysql://user3:must-be-used@localhost:666/test"
+        expect(subject.to_s).to eq 'mysql://user3:must-be-used@localhost:666/test'
       end
     end
 
     context 'with no pwd variable' do
-      let(:config_name) { 'TEST_NO_PWD' }
+      let(:mocked_env) do
+        {
+          url_key => 'mysql://user4@localhost:666/test',
+        }
+      end
 
       it 'returns the config url' do
         expect(subject).to be_a ::URI::Generic
-        expect(subject.to_s).to eq "mysql://user4@localhost:666/test"
+        expect(subject.to_s).to eq 'mysql://user4@localhost:666/test'
       end
     end
 
     context 'with no pwd variable and no credentials in base url' do
-      let(:config_name) { 'TEST_NO_AUTH' }
+      let(:mocked_env) do
+        {
+          url_key => 'mysql://localhost:666/test',
+        }
+      end
 
       it 'returns the config url' do
         expect(subject).to be_a ::URI::Generic
-        expect(subject.to_s).to eq "mysql://localhost:666/test"
+        expect(subject.to_s).to eq 'mysql://localhost:666/test'
       end
     end
 
     context 'with no host and no user and no pwd variable' do
-      let(:config_name) { 'TEST_NO_HOST_NO_AUTH' }
+      let(:mocked_env) do
+        {
+          url_key => 'sqlite:///db/test.db',
+        }
+      end
 
       it 'returns the config url' do
         expect(subject).to be_a ::URI::Generic
-        expect(subject.to_s).to eq "sqlite:///db/test.db"
+        expect(subject.to_s).to eq 'sqlite:///db/test.db'
       end
     end
 
     context 'with username in base url and pwd variable but no host' do
-      let(:config_name) { 'TEST_NO_HOST' }
+      let(:mocked_env) do
+        {
+          url_key => 'sqlite://user5@/db/test.db',
+          pwd_key => 'pwd',
+        }
+      end
 
       it 'computes the pwd in the base url' do
         expect(subject).to be_a ::URI::Generic
-        expect(subject.to_s).to eq "sqlite://user5:pwd@/db/test.db"
+        expect(subject.to_s).to eq 'sqlite://user5:pwd@/db/test.db'
       end
     end
 
-    # rubocop:disable RSpec/NestedGroups
     context 'with an unknown config name' do
       let(:config_name) { 'UNKNOWN_CONFIG' }
 
+      let(:mocked_env) do
+        {}
+      end
+
+      # rubocop:disable RSpec/MultipleMemoizedHelpers
       shared_examples 'with a fallback value' do
         let(:fallback_value) { 'postgresql://fallback@localhost:666/foo' }
 
@@ -121,19 +323,43 @@ RSpec.describe GetaroundUtils::Utils::ConfigUrl do
 
         it 'returns the fallback value' do
           expect(subject).to be_a ::URI::Generic
-          expect(subject.to_s).to eq "postgresql://fallback@localhost:666/foo"
+          expect(subject.to_s).to eq 'postgresql://fallback@localhost:666/foo'
         end
 
         it_behaves_like 'with nil fallback'
 
+        context 'with a usr variable matching' do
+          let(:mocked_env) do
+            super().merge(usr_key => 'user42')
+          end
+
+          it 'overrides the usr in the fallback value' do
+            expect(subject).to be_a ::URI::Generic
+            expect(subject.to_s).to eq 'postgresql://user42@localhost:666/foo'
+          end
+
+          it_behaves_like 'with nil fallback'
+
+          context 'with blank pwd in the fallback value' do
+            let(:fallback_value) { 'postgresql://fallback:@localhost:666/foo' }
+
+            it 'overrides the usr in the fallback value keeping blank pwd' do
+              expect(subject).to be_a ::URI::Generic
+              expect(subject.to_s).to eq 'postgresql://user42:@localhost:666/foo'
+            end
+
+            it_behaves_like 'with nil fallback'
+          end
+        end
+
         context 'with a pwd variable matching' do
-          before do
-            allow(ENV).to receive(:fetch).with('UNKNOWN_CONFIG_PASSWORD', any_args).and_return('mocked-pwd')
+          let(:mocked_env) do
+            super().merge(pwd_key => 'mocked-pwd')
           end
 
           it 'computes the pwd in the fallback value' do
             expect(subject).to be_a ::URI::Generic
-            expect(subject.to_s).to eq "postgresql://fallback:mocked-pwd@localhost:666/foo"
+            expect(subject.to_s).to eq 'postgresql://fallback:mocked-pwd@localhost:666/foo'
           end
 
           it_behaves_like 'with nil fallback'
@@ -143,13 +369,14 @@ RSpec.describe GetaroundUtils::Utils::ConfigUrl do
 
             it 'computes the pwd in the base url' do
               expect(subject).to be_a ::URI::Generic
-              expect(subject.to_s).to eq "postgresql://:mocked-pwd@localhost:666/foo"
+              expect(subject.to_s).to eq 'postgresql://:mocked-pwd@localhost:666/foo'
             end
 
             it_behaves_like 'with nil fallback'
           end
         end
       end
+      # rubocop:enable RSpec/MultipleMemoizedHelpers
 
       it { expect { subject }.to raise_error(KeyError, /UNKNOWN_CONFIG_URL/) }
 
@@ -169,6 +396,6 @@ RSpec.describe GetaroundUtils::Utils::ConfigUrl do
         it_behaves_like 'with a fallback value'
       end
     end
-    # rubocop:enable RSpec/NestedGroups
   end
+  # rubocop:enable RSpec/NestedGroups
 end